hide implementation details and update documentation
